+++ frontend/src/appComponents/axiosInstance.tsx
@@ -7,4 +7,3 @@ const axiosInstance = axios.create({
 });
 
 export default axiosInstance;
-
